requirements.txt
Flask==3.0.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
openpyxl==3.1.2
